refactor: move persistence functions to dedicated package
